requirements.txt
opencv-python
numpy
pytesseract
imutils
streamlit
Pillow
